initialmeterreading decimal value fixed

+++ web/rainmaker/dev-packages/egov-wns-dev/src/ui-config/screens/specs/wns/search-preview.js
@@ -564,7 +564,7 @@ const parserFunction = (obj) => {
       initialMeterReading: (
         obj.additionalDetails !== undefined &&
         obj.additionalDetails.initialMeterReading !== undefined
-      ) ? parseInt(obj.additionalDetails.initialMeterReading) : null,
+      ) ? parseFloat(obj.additionalDetails.initialMeterReading) : null,
       detailsProvidedBy: (
         obj.additionalDetails !== undefined &&
         obj.additionalDetails.detailsProvidedBy !== undefined &&
